+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added support for rendering 3D volume textures.
 * Added support for rendering individual models to xc3_viewer with the `--root`, `--models` and `--model` args.
+* Added `annotate-legacy-shaders` command to xc3_shader. This functionality was previously part of the `disassemble-legacy-shaders` command.
 
 ### Changed
 * Changed xc3_wgpu_batch to render wismhd models to individual files.